+++ demos/demo_5.cpp
@@ -20,7 +20,7 @@ int main(int argc, char** argv)
   audio::AudioSourceHandler src_handler;
   audio::WaveformGeneration wave_gen;
   
-  float sample_rate = 441100.f; // #FIXME: Convert to int.
+  int sample_rate = 441100;
 
   auto wd = wave_gen.generate_waveform(audio::WaveformType::SQUARE_WAVE, 3.f, 500.f,
                                       audio::FrequencyType::CONSTANT, audio::AmplitudeType::CONSTANT, audio::PhaseType::ZERO, sample_rate);
